# [REVIEW] Remove compile penalty during import