--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -10,6 +10,7 @@
 - PR #40 - Ability to specify time/freq domain for resample.
 - PR #45 - Refactor `_signaltools.py` to use new Numba/CuPy framework
 - PR #50 - Update README to reorganize install instructions
+- PR #56 - Ability to precompile select Numba/CuPy kernel on import
 
 ## Bug Fixes
 - PR #44 - Fix issues in pytests