--- a/python/cusignal/_signaltools.py
+++ b/python/cusignal/_signaltools.py
@@ -33,6 +33,13 @@ from string import Template
 from .fir_filter_design import firwin
 
 
+class GPUKernel(Enum):
+    CORRELATE = 0
+    CONVOLVE = 1
+    CORRELATE2D = 2
+    CONVOLVE2D = 3
+
+
 class GPUBackend(Enum):
     CUPY = 0
     NUMBA = 1
@@ -374,9 +381,142 @@ extern "C" {
             out[oPixelPos.x * outW + oPixelPos.y] = temp;
         }
     }
+
+    __global__ void _cupy_correlate(
+            const ${datatype} * __restrict__ inp,
+            const int inpW,
+            const ${datatype} * __restrict__ kernel,
+            const int kerW,
+            const int mode,
+            const bool swapped_inputs,
+            ${datatype} * __restrict__ out,
+            const int outW) {
+
+        const int tx {
+            static_cast<int>( blockIdx.x * blockDim.x + threadIdx.x ) };
+        const int stride { static_cast<int>( blockDim.x * gridDim.x ) };
+
+        for ( int tid = tx; tid < outW; tid += stride ) {
+            ${datatype} temp {};
+
+            if ( mode == 0 ) {  // Valid
+                if ( tid >= 0 && tid < inpW ) {
+                    for ( int j = 0; j < kerW; j++ ) {
+                        temp += inp[tid + j] * kernel[j];
+                    }
+                }
+            } else if ( mode == 1 ) {   // Same
+                const int P1 { kerW / 2 };
+                int start {};
+                if ( !swapped_inputs ) {
+                    start = 0 - P1 + tid;
+                } else {
+                    start = ( ( inpW - 1 ) / 2 ) - ( kerW - 1 ) + tid;
+                }
+                for ( int j = 0; j < kerW; j++ ) {
+                    if ( ( start + j >= 0 ) && ( start + j < inpW ) ) {
+                        temp += inp[start + j] * kernel[j];
+                    }
+                }
+            } else {    // Full
+                const int P1 { kerW - 1 };
+                int start { 0 - P1 + tid };
+                for ( int j = 0; j < kerW; j++ ) {
+                    if ( ( start + j >= 0 ) && ( start + j < inpW ) ) {
+                        temp += inp[start + j] * kernel[j];
+                    }
+                }
+            }
+
+            if (swapped_inputs) {
+                out[outW - tid - 1] = temp; // TODO: Move to shared memory
+            } else {
+                out[tid] = temp;
+            }
+        }
+    }
+    __global__ void _cupy_convolve(
+            const ${datatype} * __restrict__ inp,
+            const int inpW,
+            const ${datatype} * __restrict__ kernel,
+            const int kerW,
+            const int mode,
+            const bool swapped_inputs,
+            ${datatype} * __restrict__ out,
+            const int outW) {
+
+        const int tx {
+            static_cast<int>( blockIdx.x * blockDim.x + threadIdx.x ) };
+        const int stride { static_cast<int>( blockDim.x * gridDim.x ) };
+
+        for ( int tid = tx; tid < outW; tid += stride ) {
+            ${datatype} temp {};
+
+            if ( mode == 0 ) {  // Valid
+                if ( tid >= 0 && tid < inpW ) {
+                    for ( int j = 0; j < kerW; j++ ) {
+                        temp += inp[tid + j] * kernel[( kerW - 1 ) - j];
+                    }
+                }
+            } else if ( mode == 1 ) {   // Same
+                const int P1 { kerW / 2 };
+                int start {};
+                if ( !swapped_inputs ) {
+                    start = 0 - P1 + tid;
+                } else {
+                    start = ( ( inpW - 1 ) / 2 ) - ( kerW - 1 ) + tid;
+                }
+                for ( int j = 0; j < kerW; j++ ) {
+                    if ( ( start + j >= 0 ) && ( start + j < inpW ) ) {
+                        temp += inp[start + j] * kernel[( kerW - 1 ) - j];
+                    }
+                }
+            } else {    // Full
+                const int P1 { kerW - 1 };
+                int start { 0 - P1 + tid };
+                for ( int j = 0; j < kerW; j++ ) {
+                    if ( ( start + j >= 0 ) && ( start + j < inpW ) ) {
+                        temp += inp[start + j] * kernel[( kerW - 1 ) - j];
+                    }
+                }
+            }
+            out[tid] = temp;
+        }
+    }
 }
 """
 )
+
+
+class _cupy_convolve_wrapper(object):
+    def __init__(self, grid, block, stream, kernel):
+        if isinstance(grid, int):
+            grid = (grid,)
+        if isinstance(block, int):
+            block = (block,)
+
+        self.grid = grid
+        self.block = block
+        self.stream = stream
+        self.kernel = kernel
+
+    def __call__(
+        self, d_inp, d_kernel, mode, swapped_inputs, out,
+    ):
+
+        kernel_args = (
+            d_inp,
+            d_inp.shape[0],
+            d_kernel,
+            d_kernel.shape[0],
+            mode,
+            swapped_inputs,
+            out,
+            out.shape[0],
+        )
+
+        self.stream.use()
+        self.kernel(self.grid, self.block, kernel_args)
 
 
 class _cupy_convolve_2d_wrapper(object):
@@ -414,25 +554,38 @@ class _cupy_convolve_2d_wrapper(object):
         self.kernel(self.grid, self.block, kernel_args)
 
 
-def _get_backend_kernel(flip, dtype, grid, block, stream, use_numba):
+def _get_backend_kernel(
+    dtype, grid, block, stream, use_numba, k_type,
+):
 
-    if use_numba:
+    if not use_numba:
+        kernel = _cupy_kernel_cache[(dtype.name, k_type)]
+        if kernel:
+            if k_type == GPUKernel.CONVOLVE or k_type == GPUKernel.CORRELATE:
+                return _cupy_convolve_wrapper(grid, block, stream, kernel)
+            elif (
+                k_type == GPUKernel.CONVOLVE2D
+                or k_type == GPUKernel.CORRELATE2D
+            ):
+                return _cupy_convolve_2d_wrapper(grid, block, stream, kernel)
+            else:
+                raise Exception("If you see this let the developers know.")
+
+    else:
         nb_stream = stream_cupy_to_numba(stream)
-        kernel = _numba_kernel_cache[(flip, dtype.name)]
+        kernel = _numba_kernel_cache[(dtype.name, k_type)]
 
         if kernel:
             return kernel[grid, block, nb_stream]
-    else:
-        kernel = _cupy_kernel_cache[(flip, dtype.name)]
-        if kernel:
-            return _cupy_convolve_2d_wrapper(grid, block, stream, kernel)
 
     raise NotImplementedError(
-        "No kernel found for flip {}, datatype {}".format(flip, dtype.name)
+        "No kernel found for k_type {}, datatype {}".format(k_type, dtype.name)
     )
 
 
-def _populate_kernel_cache(np_type, use_numba):
+def _populate_kernel_cache(np_type, use_numba, k_type):
+
+    print(k_type, k_type.value)
 
     try:
         numba_type, c_type = _SUPPORTED_TYPES[np_type]
@@ -441,7 +594,7 @@ def _populate_kernel_cache(np_type, use_numba):
         raise Exception("No kernel found for datatype {}".format(np_type))
 
     if not use_numba:
-        if (0, str(numba_type)) in _cupy_kernel_cache:  # Not work
+        if (str(numba_type), k_type) in _cupy_kernel_cache:
             return
         # Instantiate the cupy kernel for this type and compile
         if isinstance(numba_type, Complex):
@@ -452,27 +605,46 @@ def _populate_kernel_cache(np_type, use_numba):
         module = cp.RawModule(
             code=src, options=("-std=c++11", "-use_fast_math")
         )
-        _cupy_kernel_cache[(0, str(numba_type))] = module.get_function(
-            "_cupy_correlate_2d"
-        )
-        _cupy_kernel_cache[(1, str(numba_type))] = module.get_function(
-            "_cupy_convolve_2d"
-        )
+        if k_type == GPUKernel.CORRELATE:
+            _cupy_kernel_cache[
+                (str(numba_type), GPUKernel.CORRELATE)
+            ] = module.get_function("_cupy_correlate")
+        elif k_type == GPUKernel.CONVOLVE:
+            _cupy_kernel_cache[
+                (str(numba_type), GPUKernel.CONVOLVE)
+            ] = module.get_function("_cupy_convolve")
+        elif k_type == GPUKernel.CORRELATE2D:
+            _cupy_kernel_cache[
+                (str(numba_type), GPUKernel.CORRELATE2D)
+            ] = module.get_function("_cupy_correlate_2d")
+        elif k_type == GPUKernel.CONVOLVE2D:
+            _cupy_kernel_cache[
+                (str(numba_type), GPUKernel.CONVOLVE2D)
+            ] = module.get_function("_cupy_convolve_2d")
+        else:
+            raise Exception("If you see this let the developers know.")
 
     else:
-        if (0, str(numba_type)) in _numba_kernel_cache:
+        if (str(numba_type), k_type) in _numba_kernel_cache:
             return
-        # JIT compile the numba kernels, flip = 0/1 (correlate/convolve)
+        # JIT compile the numba kernels
+        # k_type = 0/1 (correlate/convolve)
         sig = _numba_convolve_2d_signature(numba_type)
-        _numba_kernel_cache[(0, str(numba_type))] = cuda.jit(
-            sig, fastmath=True
-        )(_numba_correlate_2d)
-        _numba_kernel_cache[(1, str(numba_type))] = cuda.jit(
-            sig, fastmath=True
-        )(_numba_convolve_2d)
+        if k_type == GPUKernel.CORRELATE2D:
+            _numba_kernel_cache[(str(numba_type), k_type)] = cuda.jit(
+                sig, fastmath=True
+            )(_numba_correlate_2d)
+        elif k_type == GPUKernel.CONVOLVE2D:
+            _numba_kernel_cache[(str(numba_type), k_type)] = cuda.jit(
+                sig, fastmath=True
+            )(_numba_convolve_2d)
+        elif k_type == GPUKernel.CONVOLVE or k_type == GPUKernel.CORRELATE:
+            return
+        else:
+            raise Exception("If you see this let the developers know.")
 
 
-def precompile_kernels(dtype=None, backend=None):
+def precompile_kernels(dtype=None, backend=None, k_type=None):
     r"""
     Precompile GPU kernels for later use.
 
@@ -487,12 +659,60 @@ def precompile_kernels(dtype=None, backend=None):
     """
     dtype = list(dtype) if dtype else _SUPPORTED_TYPES.keys()
     backend = list(backend) if backend else list(GPUBackend)
-    for d, b in itertools.product(dtype, backend):
-        _populate_kernel_cache(d, b.value)
+    k_type = list(k_type) if k_type else list(GPUKernel)
+    for d, b, k in itertools.product(dtype, backend, k_type):
+        _populate_kernel_cache(d, b.value, k)
+
+
+def _convolve_gpu(
+    inp, out, ker, mode, use_convolve, swapped_inputs, cp_stream,
+):
+
+    d_inp = cp.array(inp)
+    d_kernel = cp.array(ker)
+
+    device_id = cp.cuda.Device()
+    numSM = device_id.attributes["MultiProcessorCount"]
+
+    threadsperblock = 256
+    blockspergrid = numSM * 20
+
+    if use_convolve:
+        _populate_kernel_cache(out.dtype.type, False, GPUKernel.CONVOLVE)
+        kernel = _get_backend_kernel(
+            out.dtype,
+            blockspergrid,
+            threadsperblock,
+            cp_stream,
+            False,
+            GPUKernel.CONVOLVE,
+        )
+    else:
+        _populate_kernel_cache(out.dtype.type, False, GPUKernel.CORRELATE)
+        kernel = _get_backend_kernel(
+            out.dtype,
+            blockspergrid,
+            threadsperblock,
+            cp_stream,
+            False,
+            GPUKernel.CORRELATE,
+        )
+
+    kernel(d_inp, d_kernel, mode, swapped_inputs, out)
+
+    return out
 
 
 def _convolve2d_gpu(
-    inp, out, kernel, mode, boundary, flip, fillvalue, cp_stream, use_numba,
+    inp,
+    out,
+    ker,
+    mode,
+    boundary,
+    use_convolve,
+    fillvalue,
+    cp_stream,
+    use_numba,
 ):
 
     if (boundary != PAD) and (boundary != REFLECT) and (boundary != CIRCULAR):
@@ -501,36 +721,36 @@ def _convolve2d_gpu(
     S = np.zeros(2, dtype=int)
 
     # If kernel is square and odd
-    if kernel.shape[0] == kernel.shape[1]:  # square
-        if kernel.shape[0] % 2 == 1:  # odd
+    if ker.shape[0] == ker.shape[1]:  # square
+        if ker.shape[0] % 2 == 1:  # odd
             pick = 1
-            S[0] = (kernel.shape[0] - 1) // 2
+            S[0] = (ker.shape[0] - 1) // 2
             if mode == 2:  # full
                 P1 = P2 = P3 = P4 = S[0] * 2
             else:  # same/valid
                 P1 = P2 = P3 = P4 = S[0]
         else:  # even
             pick = 2
-            S[0] = kernel.shape[0] // 2
+            S[0] = ker.shape[0] // 2
             if mode == 2:  # full
                 P1 = P2 = P3 = P4 = S[0] * 2 - 1
             else:  # same/valid
-                if flip:
+                if use_convolve:
                     P1 = P2 = P3 = P4 = S[0]
                 else:
                     P1 = P3 = S[0] - 1
                     P2 = P4 = S[0]
     else:  # Non-square
         pick = 3
-        S[0] = kernel.shape[0]
-        S[1] = kernel.shape[1]
+        S[0] = ker.shape[0]
+        S[1] = ker.shape[1]
         if mode == 2:  # full
             P1 = S[0] - 1
             P2 = S[0] - 1
             P3 = S[1] - 1
             P4 = S[1] - 1
         else:  # same/valid
-            if flip:
+            if use_convolve:
                 P1 = S[0] // 2
                 P2 = S[0] // 2 if (S[0] % 2) else S[0] // 2 - 1
                 P3 = S[1] // 2
@@ -566,7 +786,7 @@ def _convolve2d_gpu(
     outH = out.shape[0]
 
     d_inp = cp.array(inp)
-    d_kernel = cp.array(kernel)
+    d_kernel = cp.array(ker)
 
     threadsperblock = (16, 16)
     blockspergrid = (
@@ -574,22 +794,91 @@ def _convolve2d_gpu(
         _iDivUp(outH, threadsperblock[1]),
     )
 
-    _populate_kernel_cache(out.dtype.type, use_numba)
+    if use_convolve:
+        _populate_kernel_cache(out.dtype.type, use_numba, GPUKernel.CONVOLVE2D)
+        kernel = _get_backend_kernel(
+            out.dtype,
+            blockspergrid,
+            threadsperblock,
+            cp_stream,
+            use_numba,
+            GPUKernel.CONVOLVE2D,
+        )
+    else:
+        _populate_kernel_cache(
+            out.dtype.type, use_numba, GPUKernel.CORRELATE2D
+        )
+        kernel = _get_backend_kernel(
+            out.dtype,
+            blockspergrid,
+            threadsperblock,
+            cp_stream,
+            use_numba,
+            GPUKernel.CORRELATE2D,
+        )
 
-    kernell = _get_backend_kernel(
-        flip, out.dtype, blockspergrid, threadsperblock, cp_stream, use_numba,
-    )
-
-    kernell(
+    kernel(
         d_inp, paddedW, paddedH, d_kernel, S[0], S[1], out, outW, outH, pick
     )
+    return out
+
+
+def _convolve(
+    in1,
+    in2,
+    use_convolve,
+    swapped_inputs,
+    mode,
+    cp_stream=cp.cuda.stream.Stream(null=True),
+):
+
+    val = _valfrommode(mode)
+
+    # Promote inputs
+    promType = cp.promote_types(in1.dtype, in2.dtype)
+    in1 = in1.astype(promType)
+    in2 = in2.astype(promType)
+
+    # Create empty array to hold number of aout dimensions
+    out_dimens = np.empty(in1.ndim, np.int)
+    if val == VALID:
+        for i in range(in1.ndim):
+            out_dimens[i] = (
+                max(in1.shape[i], in2.shape[i])
+                - min(in1.shape[i], in2.shape[i])
+                + 1
+            )
+            if out_dimens[i] < 0:
+                raise Exception(
+                    "no part of the output is valid, use option 1 (same) or 2 \
+                     (full) for third argument"
+                )
+    elif val == SAME:
+        for i in range(in1.ndim):
+            if not swapped_inputs:
+                out_dimens[i] = in1.shape[i]  # Per scipy docs
+            else:
+                out_dimens[i] = min(in1.shape[i], in2.shape[i])
+    elif val == FULL:
+        for i in range(in1.ndim):
+            out_dimens[i] = in1.shape[i] + in2.shape[i] - 1
+    else:
+        raise Exception("mode must be 0 (valid), 1 (same), or 2 (full)")
+
+    # Create empty array out on GPU
+    out = cp.empty(out_dimens.tolist(), in1.dtype)
+
+    out = _convolve_gpu(
+        in1, out, in2, val, use_convolve, swapped_inputs, cp_stream=cp_stream,
+    )
+
     return out
 
 
 def _convolve2d(
     in1,
     in2,
-    flip,
+    use_convolve,
     mode="full",
     boundary="fill",
     fillvalue=0,
@@ -597,15 +886,18 @@ def _convolve2d(
     use_numba=False,
 ):
 
+    val = _valfrommode(mode)
+    bval = _bvalfromboundary(boundary)
+
     # Promote inputs
     promType = cp.promote_types(in1.dtype, in2.dtype)
     in1 = in1.astype(promType)
     in2 = in2.astype(promType)
 
-    if (boundary != PAD) and (boundary != REFLECT) and (boundary != CIRCULAR):
+    if (bval != PAD) and (bval != REFLECT) and (bval != CIRCULAR):
         raise Exception("Incorrect boundary value.")
 
-    if (boundary == PAD) and (fillvalue is not None):
+    if (bval == PAD) and (fillvalue is not None):
         fill = np.array(fillvalue, in1.dtype)
         if fill is None:
             raise Exception("If you see this let developers know.")
@@ -622,7 +914,7 @@ def _convolve2d(
 
     # Create empty array to hold number of aout dimensions
     out_dimens = np.empty(in1.ndim, np.int)
-    if mode == VALID:
+    if val == VALID:
         for i in range(in1.ndim):
             out_dimens[i] = in1.shape[i] - in2.shape[i] + 1
             if out_dimens[i] < 0:
@@ -630,10 +922,10 @@ def _convolve2d(
                     "no part of the output is valid, use option 1 (same) or 2 \
                      (full) for third argument"
                 )
-    elif mode == SAME:
+    elif val == SAME:
         for i in range(in1.ndim):
             out_dimens[i] = in1.shape[i]
-    elif mode == FULL:
+    elif val == FULL:
         for i in range(in1.ndim):
             out_dimens[i] = in1.shape[i] + in2.shape[i] - 1
     else:
@@ -646,9 +938,9 @@ def _convolve2d(
         in1,
         out,
         in2,
-        mode,
-        boundary,
-        flip,
+        val,
+        bval,
+        use_convolve,
         fill,
         cp_stream=cp_stream,
         use_numba=use_numba,