--- a/python/cusignal/_signaltools.py
+++ b/python/cusignal/_signaltools.py
@@ -662,9 +662,27 @@ def precompile_kernels(dtype=None, backend=None, k_type=None):
     dtype : numpy datatype or list of datatypes, optional
         Data types for which kernels should be precompiled. If not
         specified, all supported data types will be precompiled.
+        Specific to this unit
+            np.int32
+            np.int64
+            np.float32
+            np.float64
+            np.complex64
+            np.complex128
     backend : GPUBackend, optional
         Which GPU backend to precompile for. If not specified,
         all supported backends will be precompiled.
+        Specific to this unit
+            GPUBackend.CUPY
+            GPUBackend.NUMBA
+    k_type : GPUKernel, optional
+        Which GPU kernel to compile for. If not specified,
+        all supported kernels will be precompiled.
+        Specific to this unit
+            GPUBackend.CORRELATE
+            GPUBackend.CONVOLVE
+            GPUBackend.CORRELATE2D
+            GPUBackend.CONVOLVE2D
     """
     dtype = list(dtype) if dtype else _SUPPORTED_TYPES.keys()
     backend = list(backend) if backend else list(GPUBackend)