… python/cusignal/_upfirdn.py
… python/cusignal/benchmark/bench_signaltools.py
… python/cusignal/signaltools.py
… python/cusignal/test/test_signaltools.py