### Info
This PR is meant to address Issue https://github.com/rapidsai/cusignal/issues/54, where import time for cusignal is excessive long because all Numba and CuPy kernel are being compile during import.

This PR attempts to solve this issue by allowing users the ability to explicitly precompile desired Numba or CuPy kernel before usage and also allow kernel to be compile on first use.

Currently, this PR focuses on **cusignal.lombscargle** in `_spectral.py` as a testbed. 

1. Remove **_populate_kernel_cache()** from being called on import and move it to **_lombscargle**. This will allow kernels of a specific _dtype_ and _use_numba_ to be compiled on first use. **_populate_kernel_cache()** will now verify if a kernel has already been compiled. If yes, return; if no, compile desired kernel.

```python
def _populate_kernel_cache(np_type, use_numba):

    try:
        numba_type, c_type = _SUPPORTED_TYPES.get(np_type)
    except TypeError:
        print("Data type {} not supported.".format(np_type))
        return
    ...
```

2. Add **_precompile_kernels()** to allow users to precompile desired kernels.
```python
import cusignal
cusignal._spectral._populate_kernel_cache(np.float64, False)
```

### Results
If this technique is applied too Numba/CuPy kernels it reduces import time by 9x on my x86 machine
Before:
```bash
(cusignal) belt@orion:~/workStuff/rapids/cusignal/python$ time python3 -c "import cusignal"

real    0m4.695s
user    0m4.192s
sys     0m0.190s
```
After
```bash
(cusignal) belt@orion:~/workStuff/rapids/cusignal/python$ time python3 -c "import cusignal"

real    0m0.420s
user    0m0.363s
sys     0m0.058s
```

There is a noticeable impact on single run use cases. In this example, a 21x improvement with precompile over no-precompile.

Without precompile:
```bash
Time(%)      Total Time   Instances         Average         Minimum         Maximum  Range             
-------  --------------  ----------  --------------  --------------  --------------  ------------------
  100.0        74123635           1      74123635.0        74123635        74123635  gpu_lombscargle
```

With precompile:
```bash
Time(%)      Total Time   Instances         Average         Minimum         Maximum  Range             
-------  --------------  ----------  --------------  --------------  --------------  ------------------
  100.0         3372422           1       3372422.0         3372422         3372422  gpu_lombscargle
```

There is no visible impact on benchmarks, see below
```bash
Name (time in us)                                             Min                    Max                  Mean                StdDev                Median                IQR            Outliers       OPS            Rounds  Iterations
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
bench_lombscargle_gpu[False-False-False-262144-1024]      26.9060 (1.0)      56,899.5480 (2.81)     4,287.2436 (3.87)     1,424.2830 (2.38)     4,348.9110 (3.87)      3.0130 (2.29)    1414;5887  233.2501 (0.26)      35405           1
bench_lombscargle_gpu[False-False-False-65536-1024]       27.0030 (1.00)     39,603.8470 (1.95)     1,106.9343 (1.0)        677.2651 (1.13)     1,123.0970 (1.0)       1.3130 (1.0)     1509;7552  903.3960 (1.0)       35722           1
bench_lombscargle_gpu[False-False-True-262144-1024]       49.8970 (1.85)     64,352.6260 (3.18)     4,867.9718 (4.40)     1,592.3355 (2.66)     4,928.2250 (4.39)     10.3015 (7.85)     767;2721  205.4244 (0.23)      19584           1
bench_lombscargle_gpu[False-True-False-65536-1024]        50.6500 (1.88)     43,811.1860 (2.16)     1,121.7380 (1.01)       717.9696 (1.20)     1,137.9840 (1.01)     40.5930 (30.92)    949;1299  891.4738 (0.99)      19430           1
bench_lombscargle_gpu[False-True-False-262144-1024]       50.9030 (1.89)     54,865.4490 (2.71)     4,733.9890 (4.28)     1,630.7039 (2.73)     4,838.5610 (4.31)     11.4735 (8.74)     851;5062  211.2383 (0.23)      19219           1
bench_lombscargle_gpu[False-False-True-65536-1024]        51.5440 (1.92)     46,816.6060 (2.31)     1,303.1182 (1.18)       848.3968 (1.42)     1,319.8260 (1.18)      3.5898 (2.73)     823;4650  767.3901 (0.85)      19023           1
bench_lombscargle_gpu[False-True-True-65536-1024]         70.0490 (2.60)     26,646.1420 (1.31)     1,387.8114 (1.25)       658.6363 (1.10)     1,406.3755 (1.25)     12.5560 (9.56)     626;2107  720.5590 (0.80)      14066           1
bench_lombscargle_gpu[False-True-True-262144-1024]        71.3590 (2.65)     48,045.3240 (2.37)     4,519.1759 (4.08)     1,503.1672 (2.51)     4,565.3610 (4.06)     70.9170 (54.01)    578;1556  221.2793 (0.24)      13711           1
bench_lombscargle_gpu[True-False-False-262144-1024]      285.0280 (10.59)    28,164.6460 (1.39)     4,243.6270 (3.83)     1,879.3081 (3.14)     4,944.0910 (4.40)     13.8915 (10.58)     577;732  235.6475 (0.26)       3468           1
bench_lombscargle_gpu[True-False-False-65536-1024]       291.9050 (10.85)    21,443.2310 (1.06)     1,164.8023 (1.05)       668.6357 (1.12)     1,361.3015 (1.21)     16.5910 (12.64)     663;802  858.5148 (0.95)       3382           1
bench_lombscargle_gpu[True-True-False-262144-1024]       312.0630 (11.60)    43,368.4600 (2.14)     4,459.1994 (4.03)     1,957.5219 (3.27)     4,958.3820 (4.41)     50.6218 (38.55)     425;659  224.2555 (0.25)       3169           1
bench_lombscargle_gpu[True-True-False-65536-1024]        312.4470 (11.61)    21,014.0490 (1.04)     1,225.7864 (1.11)       623.5198 (1.04)     1,363.5020 (1.21)     31.7132 (24.15)     498;824  815.8028 (0.90)       3173           1
bench_lombscargle_gpu[True-False-True-65536-1024]        323.6380 (12.03)    22,066.5290 (1.09)     1,244.0159 (1.12)       615.7444 (1.03)     1,349.6860 (1.20)     20.1062 (15.31)     379;662  803.8483 (0.89)       3023           1
bench_lombscargle_gpu[True-False-True-262144-1024]       325.8770 (12.11)    35,460.8630 (1.75)     4,050.7925 (3.66)     1,493.7356 (2.50)     4,388.9525 (3.91)     43.5940 (33.20)     325;564  246.8653 (0.27)       3054           1
bench_lombscargle_gpu[True-True-True-262144-1024]        339.3310 (12.61)    28,899.2560 (1.43)     4,503.7586 (4.07)     1,387.3032 (2.32)     4,807.5840 (4.28)     69.5670 (52.98)     251;474  222.0368 (0.25)       2902           1
bench_lombscargle_gpu[True-True-True-65536-1024]         345.1900 (12.83)    20,264.4000 (1.0)      1,310.9682 (1.18)       598.4111 (1.0)      1,384.0260 (1.23)     20.9048 (15.92)     367;718  762.7950 (0.84)       2811           1
```

If there are no issues, I'll apply these changes to the other kernels and submit next week.